corrections for adding primitive data type

+++ Assignment 2/src/ui/ViewSearchJPanel.java
@@ -501,7 +501,7 @@ public class ViewSearchJPanel extends javax.swing.JPanel {
         foundPerson.setAge(txtAge.getText().getBytes()[0]);
         foundPerson.setLastName(txtLastName.getText());
         foundPerson.setSsn(txtSsn.getText());
-        foundPerson.setHomeAddress(txtStreetH.getText(),txtStateH.getText(),Integer.valueOf(txtZipH.getText()),txtCityH.getText(),Integer.valueOf(txtUnitH.getText()),Long.valueOf(txtPhoneH.getText()));
+        foundPerson.setHomeAddress(txtStreetH.getText(),txtStateH.getText(),Integer.valueOf(txtZipH.getText()),txtCityH.getText(),Short.valueOf(txtUnitH.getText()),Long.valueOf(txtPhoneH.getText()));
         foundPerson.setWorkAddress(txtStreetW.getText(),txtStateW.getText(),Integer.valueOf(txtZipW.getText()),txtCityW.getText(),Integer.valueOf(txtUnitW.getText()),Long.valueOf(txtPhoneW.getText()));
